🚀 Feat: Present Query 위치 변경
- Code Split 을 진행 하면서 데이터를 각각의 페이지에서 Fetching 하도록 수정
- Present 페이지를 방문 하지않으면 헤더 부분의 진행시간을 나타낼수 없는 문제
- Present 데이터를 캐싱할 수 있도록 레이아웃 단에서 Fetching 하도록 수정
